backend/payment-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymongo==4.6.0
motor==3.3.2
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0
python-multipart==0.0.6
